feat: poll result ui
