+++ src/v4/component/CheckBox/index.tsx
@@ -53,6 +53,7 @@ type OptionProps<T extends any = string | number> = PropsWithChildren<
   style?: (state: {
     pressed: boolean;
     selected: boolean;
+    disabled: boolean;
   }) => StyleProp<ViewStyle>;
   disabled?: boolean;
   value: T;
@@ -117,7 +118,7 @@ function Option<T extends any = string | number>({
         style={({ pressed }) => [
           styles.option,
           pressed && styles.optionPressed,
-          style && style({ pressed, selected }),
+          style && style({ pressed, selected, disabled: isDisabled }),
         ]}
       >
         {children}
